# lab6/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Comment API
  description: API для управления комментариями
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Локальный сервер

paths:
  /comments:
    post:
      summary: Создать новый комментарий
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
        '201':
          description: Комментарий успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'

  /comments/{id}:
    get:
      summary: Получить комментарий по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description: Комментарий не найден
    put:
      summary: Обновить комментарий по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
        '200':
          description: Комментарий успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          description: Комментарий не найден
    delete:
      summary: Удалить комментарий по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Комментарий успешно удален
        '404':
          description: Комментарий не найден

components:
  schemas:
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
          description: Содержимое комментария
        author:
          type: string
          description: Автор комментария
      required:
        - content
        - author